+++ frontend/src/components/Trophy/LongestStreakCountTrophy.tsx
